requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
pydantic==2.8.2
requests==2.32.3
python-dotenv==1.0.1

